fix(review): move TestSessionIndexRoundtrip to tests/execution/ to fix cross-layer import
The core-layer test file imported _flush from tests.execution.conftest,
violating the layer boundary. Moved the roundtrip test class to
tests/execution/test_session_index_roundtrip.py with proper layer marker.
